+++ WordleTests/MainTests.cs
@@ -19,22 +19,22 @@ namespace WordleTests
 			Search search = new Search(results, 0);
 			Assert.AreEqual("sanes", search.Run());
 
-			List<(char, short)> feedback = new List<(char, short)> { ('s', 0), ('a', 1), ('n', 0), ('e', 1), ('s', 0) };
+			List<(char, byte)> feedback = new List<(char, byte)> { ('s', 0), ('a', 1), ('n', 0), ('e', 1), ('s', 0) };
 			results = analyzer.Run(feedback);
 			search = new Search(results, 0);
 			Assert.AreEqual("areae", search.Run());
 
-			feedback = new List<(char, short)> { ('a', 1), ('r', 0), ('e', 1), ('a', 1), ('e', 1) };
+			feedback = new List<(char, byte)> { ('a', 1), ('r', 0), ('e', 1), ('a', 1), ('e', 1) };
 			results = analyzer.Run(feedback);
 			search = new Search(results, 0);
 			Assert.AreEqual("beata", search.Run());
 
-			feedback = new List<(char, short)> { ('b', 0), ('e', 2), ('a', 2), ('t', 2), ('a', 1) };
+			feedback = new List<(char, byte)> { ('b', 0), ('e', 2), ('a', 2), ('t', 2), ('a', 1) };
 			results = analyzer.Run(feedback);
 			search = new Search(results, 0);
 			Assert.AreEqual("meath", search.Run());
 
-			feedback = new List<(char, short)> { ('m', 2), ('e', 2), ('a', 2), ('t', 2), ('h', 0) };
+			feedback = new List<(char, byte)> { ('m', 2), ('e', 2), ('a', 2), ('t', 2), ('h', 0) };
 			results = analyzer.Run(feedback);
 			search = new Search(results, 0);
 			Assert.AreEqual("meaty", search.Run());
@@ -47,22 +47,22 @@ namespace WordleTests
 			Search search = new Search(results);
 			Assert.AreEqual("cares", search.Run());
 
-			List<(char, short)> feedback = new List<(char, short)> { ('c', 0), ('a', 1), ('r', 0), ('e', 1), ('s', 0) };
+			List<(char, byte)> feedback = new List<(char, byte)> { ('c', 0), ('a', 1), ('r', 0), ('e', 1), ('s', 0) };
 			results = analyzer.Run(feedback);
 			search = new Search(results);
 			Assert.AreEqual("alane", search.Run());
 
-			feedback = new List<(char, short)> { ('a', 1), ('l', 0), ('a', 2), ('n', 0), ('e', 1) };
+			feedback = new List<(char, byte)> { ('a', 1), ('l', 0), ('a', 2), ('n', 0), ('e', 1) };
 			results = analyzer.Run(feedback);
 			search = new Search(results);
 			Assert.AreEqual("peaty", search.Run());
 
-			feedback = new List<(char, short)> { ('p', 0), ('e', 2), ('a', 2), ('t', 2), ('y', 2) };
+			feedback = new List<(char, byte)> { ('p', 0), ('e', 2), ('a', 2), ('t', 2), ('y', 2) };
 			results = analyzer.Run(feedback);
 			search = new Search(results);
 			Assert.AreEqual("featy", search.Run());
 
-			feedback = new List<(char, short)> { ('f', 0), ('e', 2), ('a', 2), ('t', 2), ('y', 2) };
+			feedback = new List<(char, byte)> { ('f', 0), ('e', 2), ('a', 2), ('t', 2), ('y', 2) };
 			results = analyzer.Run(feedback);
 			search = new Search(results);
 			Assert.AreEqual("meaty", search.Run());
@@ -72,7 +72,7 @@ namespace WordleTests
 		public void AnalyzerSavedTest()
 		{
 			Analyzer analyzer = new Analyzer(5);
-			(List<List<(char, int)>>, List<(char, int)>) results = analyzer.Run(new List<(char, short)> { ('s',2),('a',0),('i',2),('n',2),('t',2) });
+			(List<List<(char, int)>>, List<(char, int)>) results = analyzer.Run(new List<(char, byte)> { ('s',2),('a',0),('i',2),('n',2),('t',2) });
 			Search search = new Search(results);
 			string first = search.Run();
 			var result2 = analyzer.Run(null);
